Add SvelteKit v1.0+ compatibility
