+++ package.json
@@ -3,33 +3,50 @@
   "license": "MIT",
   "version": "1.6.2",
   "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "svelte-kit package",
-    "package": "svelte-kit package",
-    "build:demo": "svelte-kit build",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "preview": "svelte-kit preview"
+    "dev": "vite dev",
+    "build": "vite build && yarn run package",
+    "preview": "vite preview",
+    "package": "svelte-kit sync && svelte-package && publint",
+    "prepublishOnly": "npm run package",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/adapter-cloudflare": "^1.0.0-next.9",
-    "@sveltejs/kit": "^next",
-    "svelte": "^3.46.4",
-    "svelte-check": "^2.4.2",
-    "svelte-preprocess": "^4.9.4",
-    "typescript": "^4.4.3",
-    "sass": "^1.49.7"
+    "@sveltejs/adapter-cloudflare": "^2.3.0",
+    "@sveltejs/kit": "^1.20.4",
+    "@sveltejs/package": "^2.0.0",
+    "@types/events": "^3.0.0",
+    "@types/lodash": "^4.14.195",
+    "@types/uuid": "^9.0.2",
+    "publint": "^0.1.12",
+    "sass": "^1.63.6",
+    "svelte": "^3.59.2",
+    "svelte-check": "^3.4.4",
+    "svelte-preprocess": "^5.0.4",
+    "tslib": "^2.4.1",
+    "typescript": "^5.0.0",
+    "vite": "^4.3.9"
   },
   "type": "module",
-  "main": "./index.js",
+  "svelte": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "dependencies": {
     "events": "^3.3.0",
     "lodash": "^4.17.21",
     "stampit": "^4.3.2",
-    "svelte-select": "^4.4.4",
-    "svelte2tsx": "^0.5.2",
-    "tslib": "^2.3.1",
-    "uuid": "^8.3.2",
+    "svelte-select": "^4.4.7",
+    "svelte2tsx": "^0.6.16",
+    "uuid": "^9.0.0",
     "visibilityjs": "^2.0.2"
+  },
+  "peerDependencies": {
+    "@sveltejs/kit": "^1.20.4",
+    "svelte": "^3.59.2"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "svelte": "./dist/index.js"
+    }
   }
 }
